discord_bot_api/requirements.txt
discord.py==2.3.2
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
